requirements.txt
Flask
flask-cors
python-dotenv
psycopg2-binary
APScheduler
imagehash
Pillow
requests
PyJWT